seperated constants into ther own file in a constants folder
