+++ src/main/java/frc/robot/constants/IntakeConstants.java
@@ -1,0 +1,7 @@
+package frc.robot.constants;
+
+public class IntakeConstants {
+    public static final int intakeMotorID = 15;
+
+    public static final double intakeMotorCurrentLimit = 5;
+}
