+++ src/main/java/frc/robot/subsystems/drive/ModuleIOSim.java
@@ -13,7 +13,7 @@
 
 package frc.robot.subsystems.drive;
 
-import static frc.robot.Constants.DriveConstants.Module.WHEEL_RADIUS;
+import static frc.robot.constants.DriveConstants.Module.WHEEL_RADIUS;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.PIDController;
