+++ src/main/java/frc/robot/subsystems/drive/PhoenixOdometryThread.java
@@ -18,7 +18,7 @@ import com.ctre.phoenix6.CANBus;
 import com.ctre.phoenix6.StatusSignal;
 import com.ctre.phoenix6.hardware.ParentDevice;
 import edu.wpi.first.units.measure.Angle;
-import frc.robot.Constants.DriveConstants;
+import frc.robot.constants.DriveConstants;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.List;
