+++ src/main/java/frc/robot/subsystems/drive/SparkMaxOdometryThread.java
@@ -14,7 +14,7 @@
 package frc.robot.subsystems.drive;
 
 import edu.wpi.first.wpilibj.Notifier;
-import frc.robot.Constants.DriveConstants;
+import frc.robot.constants.DriveConstants;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.List;
