+++ src/main/java/frc/robot/subsystems/intake/IntakeIOReal.java
@@ -1,6 +1,6 @@
 package frc.robot.subsystems.intake;
 
-import static frc.robot.Constants.IntakeConstants.*;
+import static frc.robot.constants.IntakeConstants.*;
 
 import com.ctre.phoenix6.configs.TalonFXConfiguration;
 import com.ctre.phoenix6.hardware.TalonFX;
