+++ src/main/java/frc/robot/subsystems/pivot/Pivot.java
@@ -2,10 +2,10 @@ package frc.robot.subsystems.pivot;
 
 import static edu.wpi.first.units.Units.Second;
 import static edu.wpi.first.units.Units.Volts;
-import static frc.robot.Constants.PivotConstants.SYS_ID_RAMP_RATE;
-import static frc.robot.Constants.PivotConstants.SYS_ID_STEP_VALUE;
-import static frc.robot.Constants.PivotConstants.SYS_ID_TIMEOUT;
-import static frc.robot.Constants.UPDATE_PERIOD;
+import static frc.robot.constants.PivotConstants.SYS_ID_RAMP_RATE;
+import static frc.robot.constants.PivotConstants.SYS_ID_STEP_VALUE;
+import static frc.robot.constants.PivotConstants.SYS_ID_TIMEOUT;
+import static frc.robot.constants.GeneralConstants.UPDATE_PERIOD;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.RunCommand;
