+++ src/main/java/frc/robot/util/PathGenerator.java
@@ -8,7 +8,7 @@ import com.pathplanner.lib.path.Waypoint;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants.PathingConstants;
+import frc.robot.constants.PathingConstants;
 import java.util.List;
 
 public class PathGenerator {
